+++ test/internal/app/mailers/child_mailer.rb
@@ -1,0 +1,3 @@
+class ChildMailer < ParentMailer
+  has_history message: false, only: [:other]
+end