+++ test/internal/app/mailers/extra_mailer.rb
@@ -1,6 +1,7 @@
 class ExtraMailer < ApplicationMailer
-  track extra: {coupon_id: 1}, only: [:basic]
-  track extra: -> { {coupon_id: @coupon_id} }, only: [:dynamic]
+  has_history
+  has_history extra: {coupon_id: 1}, only: [:basic]
+  has_history extra: -> { {coupon_id: @coupon_id} }, only: [:dynamic]
 
   def welcome
     mail