+++ test/internal/app/mailers/parent_mailer.rb
@@ -1,0 +1,11 @@
+class ParentMailer < ApplicationMailer
+  has_history
+
+  def welcome
+    mail
+  end
+
+  def other
+    mail
+  end
+end